+++ demo-web-app/src/test/java/pl/akmf/ksef/sdk/api/services/QrCodeTests.java
@@ -19,7 +19,7 @@ import static org.junit.Assert.assertTrue;
 
 class QrCodeTests {
 
-    private String apiUrl = "https://ksef-test.mf.gov.pl/";
+    private String qrUrl = "https://ksef-test.mf.gov.pl/";
 
     // =============================================
     // Testy RSA; NIEZALECANE):
@@ -31,7 +31,7 @@ class QrCodeTests {
     // =============================================
     @Test
     void buildCertificateQr_WithEmbeddedPrivateKey_ShouldReturnBase64Png() throws Exception {
-        VerificationLinkService linkSvc = new DefaultVerificationLinkService(apiUrl);
+        VerificationLinkService linkSvc = new DefaultVerificationLinkService(qrUrl);
         QrCodeService qrSvc = new DefaultQrCodeService();
         CertificateBuilders.X500NameHolder x500 = new CertificateBuilders()
                 .buildForOrganization("Kowalski sp. z o.o", "VATPL-1111111111", "Kowalski", "PL");
@@ -55,7 +55,7 @@ class QrCodeTests {
 
     @Test
     void buildCertificateQr_PublicOnlyWithoutPrivateKey_ShouldThrow() throws Exception {
-        VerificationLinkService linkSvc = new DefaultVerificationLinkService(apiUrl);
+        VerificationLinkService linkSvc = new DefaultVerificationLinkService(qrUrl);
         String nip = "0000000000";
         String xml = "<x/>";
         String serial = UUID.randomUUID().toString();
@@ -75,7 +75,7 @@ class QrCodeTests {
     // =============================================
     @Test
     void buildCertificateQr_WithEmbeddedEccPrivateKey_ShouldReturnBase64Png() throws Exception {
-        VerificationLinkService linkSvc = new DefaultVerificationLinkService(apiUrl);
+        VerificationLinkService linkSvc = new DefaultVerificationLinkService(qrUrl);
         QrCodeService qrSvc = new DefaultQrCodeService();
         CertificateBuilders.X500NameHolder x500 = new CertificateBuilders()
                 .buildForOrganization("Kowalski sp. z o.o", "VATPL-1111111111", "TestEcc", "PL");