+++ ksef-client/src/main/java/pl/akmf/ksef/sdk/api/builders/batch/OpenBatchSessionRequestBuilder.java
@@ -56,21 +56,6 @@ public class OpenBatchSessionRequestBuilder {
         return this;
     }
 
-    @Deprecated
-    public OpenBatchSessionRequestBuilder addBatchFilePart(int ordinalNumber, String fileName, long fileSize, String fileHash) {
-        if (ordinalNumber < 0 || isNullOrBlank(fileName) || fileSize < 0 || isNullOrBlank(fileHash)) {
-            throw new IllegalArgumentException("BatchFilePart parameters are invalid.");
-        }
-
-        BatchFilePartInfo batchFilePartInfo = new BatchFilePartInfo();
-        batchFilePartInfo.setOrdinalNumber(ordinalNumber);
-        batchFilePartInfo.setFileName(fileName);
-        batchFilePartInfo.setFileSize(fileSize);
-        batchFilePartInfo.setFileHash(fileHash);
-        this.parts.add(batchFilePartInfo);
-        return this;
-    }
-
     public OpenBatchSessionRequestBuilder addBatchFilePart(int ordinalNumber, long fileSize, String fileHash) {
         if (ordinalNumber < 0 || fileSize < 0 || isNullOrBlank(fileHash)) {
             throw new IllegalArgumentException("BatchFilePart parameters are invalid.");